Module difft::conflicts

Apply conflict markers to obtain the original file contents.

https://git-scm.com/docs/git-merge#Documentation/git-merge.txt-mergeconflictStyle

Structs

ConflictFiles 🔒

Enums

ConflictState 🔒

Constants

END_RHS_MARKER 🔒

START_BASE_MARKER 🔒

START_LHS_MARKER 🔒

START_RHS_MARKER 🔒

Functions

apply_conflict_markers 🔒

Convert a string with conflict markers into the two conflicting file contents.