Module difft::diff::dijkstra

Implements Dijkstra’s algorithm for shortest path, to find an optimal and readable diff between two ASTs.

Structs

ExceededGraphLimit 🔒

Functions

edge_between 🔒

mark_syntax 🔒

node_count 🔒

What is the total number of AST nodes?

shortest_path 🔒

Return the shortest route from the start to the end vertex.

shortest_path_with_edges 🔒

shortest_vertex_path 🔒

Return the shortest route from start to the end vertex.

tree_count 🔒

How many top-level AST nodes do we have?