difft::diff::dijkstra
Function shortest_path

fn shortest_path<'s, 'v>(
    start: Vertex<'s, 'v>,
    vertex_arena: &'v Bump,
    size_hint: usize,
    graph_limit: usize,
) -> Result<Vec<(Edge, &'v Vertex<'s, 'v>)>, ExceededGraphLimit>

Return the shortest route from the start to the end vertex.

The vec returned does not return the very last vertex. This is necessary because a route of N vertices only has N-1 edges.