Module difft::diff::graph

A graph representation for computing tree diffs.

Structs

Vertex

A vertex in a directed acyclic graph that represents a diff.

Enums

Edge

An edge in our graph, with an associated cost.

EnteredDelimiter

Tracks entering syntax List nodes.

Constants

NOT_CONTIGUOUS_PENALTY

Functions

allocate_if_new

can_pop_either_parent

get_set_neighbours

Compute the neighbours of v if we haven’t previously done so, write them to the .neighbours cell inside v, and return them.

looks_like_punctuation

Does this atom look like punctuation?

pop_all_parents

Pop as many parents of lhs_node and rhs_node as possible. Return the new syntax nodes and parents.

populate_change_map

push_both_delimiters

push_lhs_delimiter

push_rhs_delimiter

try_pop_both

try_pop_lhs

try_pop_rhs