Enum difft::diff::graph::Edge

pub(crate) enum Edge {
    UnchangedNode {
        depth_difference: u32,
        probably_punctuation: bool,
    },
    EnterUnchangedDelimiter {
        depth_difference: u32,
    },
    ReplacedComment {
        levenshtein_pct: u8,
    },
    ReplacedString {
        levenshtein_pct: u8,
    },
    NovelAtomLHS {},
    NovelAtomRHS {},
    EnterNovelDelimiterLHS {},
    EnterNovelDelimiterRHS {},
}

An edge in our graph, with an associated cost.

A syntax node can always be marked as novel, so a vertex will have at least a NovelFoo edge. Depending on the syntax nodes of the current Vertex, other edges may also be available.

See set_neighbours for all the edges available for a given Vertex.

Variants

UnchangedNode

Fields

depth_difference: u32

probably_punctuation: bool

Is this node just punctuation? We penalise this case, because it’s more useful to match e.g. a variable name than a comma.

EnterUnchangedDelimiter

Fields

depth_difference: u32

ReplacedComment

Fields

levenshtein_pct: u8

ReplacedString

Fields

levenshtein_pct: u8

NovelAtomLHS

NovelAtomRHS

EnterNovelDelimiterLHS

EnterNovelDelimiterRHS

Implementations

impl Edge

pub(crate) fn cost(self) -> u32