difft::diff
Module lcs_diff

A fast diff for linear content, particularly lines of text.

This file uses the Wu algorithm, using the wu-diff crate.

Difftastic has the files huge_cpp_1.cpp and huge_cpp_2.cpp in the sample_files directory for a performance stress test. These files are 22 MiB and 590,000 lines.

Enums

DiffResult 🔒

Functions

slice 🔒

Compute a linear diff between lhs and rhs.

slice_by_hash 🔒

Compute a linear diff between lhs and rhs, but use hashed values internally.

slice_unique_by_hash 🔒

Compute the linear diff between lhs and rhs. If there are items that only occur on a single side, mark them as novel without processing them with Myers’ diff.