Function difft::diff::myers_diff::slice_by_hash

pub(crate) fn slice_by_hash<'a, T: Eq + Hash>(
    lhs: &'a [T],
    rhs: &'a [T]
) -> Vec<DiffResult<&'a T>>

Compute a linear diff between lhs and rhs, but use hashed values internally.

This is faster when equality checks on T are expensive, such as large strings.