difft::display
Module context

Calculate which nearby lines should also be displayed.

Functions

add_context 🔒

add_ends 🔒

Extend matched_lines to include the leading and trailing lines in the file.

after_with_opposites 🔒

Before: 120 – (novel) 121 – 122 –

all_lines 🔒

all_matched_lines 🔒

all_matched_lines_filled 🔒

before_with_opposites 🔒

Before:

calculate_after_context 🔒

calculate_before_context 🔒

compact_gaps 🔒

Before:

ensure_contiguous 🔒

Ensure that we don’t miss any intermediate values.

flip_tuple 🔒

flip_tuples 🔒

match_blanks_before 🔒

match_blanks_between 🔒

If the lines immediately before current are blank on both sides, return the line numbers of those blank lines.

match_preceding_blanks 🔒

For every line number pair, if there are blank lines preceding the pair, add those blank lines to the vec.

matched_lines_from_unchanged 🔒

merge_in_opposite_lines 🔒

opposite_positions 🔒

pad_after 🔒

pad_before 🔒