Module difft::exit_codes

Constants

EXIT_BAD_ARGUMENTS 🔒

Invalid arguments given to difftastic. This could be usage errors (e.g. invalid numbers of arguments) or invalid paths (e.g. files we don’t have permission to read).

EXIT_FOUND_CHANGES 🔒

Successfully ran a diff, found syntactic changes in text files or byte changes in binary files.

EXIT_SUCCESS 🔒

Successfully ran a diff, found no syntactic changes in text files or byte changes in binary files.