difft
Module files

File reading utilities.

Enums

ProbableFileKind 🔒

Functions

eprint_read_error 🔒

Write a human-friendly description of e to stderr.

guess_content 🔒

Do these bytes look like a text or a binary format? If it’s text, decode it.

has_utf16_byte_order_mark 🔒

read_file_arg 🔒

Read a path provided in a CLI argument, handling /dev/null and - correctly.

read_file_or_die 🔒

read_files_or_die 🔒

read_or_die 🔒

relative_file_paths_in_dir 🔒

All the files in dir, including subdirectories.

relative_paths_in_either 🔒

Walk lhs_dir and rhs_dir, and return relative paths of files that occur in at least one directory.

u16_from_bytes 🔒

Group bytes into u16 values for conversion to UTF-16, respecting the byte order mark if present.