Function difft::diff_directories

pub(crate) fn diff_directories<'a>(
    lhs_dir: &'a Path,
    rhs_dir: &'a Path,
    display_options: &DisplayOptions,
    diff_options: &DiffOptions,
    overrides: &[(LanguageOverride, Vec<Pattern>)]
) -> impl ParallelIterator<Item = DiffResult> + 'a

Given two directories that contain the files, compare them pairwise. Returns an iterator, so we can print results incrementally.

When more than one file is modified, the hg extdiff extension passes directory paths with all the modified files.