difft
Module gitattributes

Enums

DiffAttribute 🔒

The result of checking file’s diff and binary attributes with git check-attr. See man gitattributes, specifically Generating diff text section.

Functions

check_diff_attr 🔒

Runs git check-attr diff binary to get the diff and binary attributes of the path. Returns Option::None when either git is not available, file is not inside git directory, or something else went wrong.

parse_output 🔒