difft::gitattributes
Function check_diff_attr

pub(crate) fn check_diff_attr(path: &Path) -> Option<DiffAttribute>

Runs git check-attr diff binary to get the diff and binary attributes of the path. Returns Option::None when either git is not available, file is not inside git directory, or something else went wrong.