Module difft::hunks

Calculating which modified lines should be displayed together.

Structs

Hunk
A hunk represents a series of modified lines that are displayed together.

Enums

Side

Constants

MAX_DISTANCE
The maximum number of lines that may occur between changed lines in a hunk.

Functions

aligned_lines_from_hunk

compact_gaps
Before:

enforce_increasing
Given a slice of line pairs that are ordered as far as possible, but may refer to earlier lines, return line pairs that are monotonically increasing. If an earlier line occurs after a later line, replace that line with None.

ensure_contiguous
Ensure that we don’t miss any intermediate values.

extract_lines

fill_aligned
Fill matched pairs from start to end.

fill_between

fill_matched_lines
Return all the matched lines, starting from lhs_start, up to end or the end of the whole file.

first_greater
Return the smallest item in item_set that is greater than value.

lines_are_close

lines_to_hunks
Split lines into hunks.

matched_pos_to_hunks

merge_adjacent

sorted_novel_positions
Return a vec of novel MatchedPos values in an order suited for displaying.

split_first_pair
Find the first pair in this vec where both items are Some. Return that pair, plus all the items that occur afterwards.

split_last_pair