Module difft::lines

Manipulate lines of text and groups of lines.

Structs

LineNumber
A distinct number type for line numbers, to prevent confusion with other numerical data.

LinePosition
A position in a single line of a string.

NewlinePositions
A struct for efficiently converting absolute string positions to line-relative positions.

Traits

MaxLine

Functions

codepoint_len
Return the length of s in codepoints. This is important when finding character boundaries for slicing without errors.

format_line_num

substring_by_codepoint
The first len codepoints of s. This is safer than slicing by bytes, which panics if the byte isn’t on a codepoint boundary.