difft::parse
Module tree_sitter_parser

Load and configure parsers written with tree-sitter.

Structs

ExceededByteLimit 🔒

ExceededParseErrorLimit 🔒

HighlightedNodeIds 🔒

TreeSitterConfig 🔒

Configuration for a tree-sitter parser.

TreeSitterSubLanguage 🔒

A language may contain certain nodes that are in other languages and should be parsed as such (e.g. HTML <script> nodes containing JavaScript). This struct describes how to identify such nodes, and what languages we should parse them as.

Constants

OCAML_ATOM_NODES 🔒

Functions

all_syntaxes_from_cursor 🔒

Convert all the tree-sitter nodes at this level to difftastic syntax nodes.

atom_from_cursor 🔒

Convert the tree-sitter node at cursor to a difftastic atom.

child_tokens 🔒

comment_positions 🔒

find_delim_positions 🔒

Are any of the children of the node at cursor delimiters? Return their indexes if so.

from_language 🔒

list_from_cursor 🔒

Convert the tree-sitter node at cursor to a difftastic list node.

parse 🔒

Parse src with tree-sitter and convert to difftastic Syntax.

parse_subtrees 🔒

Find any nodes that can be parsed as other languages (e.g. JavaScript embedded in HTML), and return a map of their node IDs mapped to parsed trees. Every time we see such a node, we will ignore it and recurse into the root node of the given tree instead.

print_cursor 🔒

print_tree 🔒

should_ignore_last_child 🔒

Does node match the ignorable trailing tokens configuration for this language?

syntax_from_cursor 🔒

Convert the tree-sitter node at cursor to a difftastic syntax node.

to_syntax 🔒

to_syntax_with_limit 🔒

to_tree 🔒

Parse src with tree-sitter.

to_tree_with_limit 🔒

tree_highlights 🔒

Calculate which tree-sitter node IDs should have which syntax highlighting.

tree_sitter_elvish 🔒 ⚠

tree_sitter_hack 🔒 ⚠

tree_sitter_hare 🔒 ⚠

tree_sitter_janet_simple 🔒 ⚠

tree_sitter_kotlin 🔒 ⚠

tree_sitter_latex 🔒 ⚠

tree_sitter_scss 🔒 ⚠

tree_sitter_smali 🔒 ⚠