Module difft::style

Apply colours and styling to strings.

Enums

BackgroundColor

Functions

apply

Apply the Styles to the spans specified.

apply_colors

apply_header_color

apply_line

Return a copy of line with styles applied to all the spans specified.

color_positions

group_by_line

header

highlight_missing_style_bug

novel_style

split_and_apply

Split line (from the source code) into multiple lines of max_len (i.e. word wrapping), and apply styles to each part according to its original position in line.

split_string_by_codepoint

Split a string into equal length parts, padding the last part if necessary.

substring_by_byte

substring_by_codepoint

Slice s from start inclusive to end exclusive by codepoint. This is safer than slicing by bytes, which panics if the byte isn’t on a codepoint boundary.