Enum difft::syntax::Syntax

pub enum Syntax<'a> {
    List {
        info: SyntaxInfo<'a>,
        open_position: Vec<SingleLineSpan>,
        open_content: String,
        children: Vec<&'a Syntax<'a>>,
        close_position: Vec<SingleLineSpan>,
        close_content: String,
        num_descendants: u32,
    },
    Atom {
        info: SyntaxInfo<'a>,
        position: Vec<SingleLineSpan>,
        content: String,
        kind: AtomKind,
    },
}

Variants

List

Fields of List

info: SyntaxInfo<'a>open_position: Vec<SingleLineSpan>open_content: Stringchildren: Vec<&'a Syntax<'a>>close_position: Vec<SingleLineSpan>close_content: Stringnum_descendants: u32

Atom

Fields of Atom

info: SyntaxInfo<'a>position: Vec<SingleLineSpan>content: Stringkind: AtomKind

Implementations

impl<'a> Syntax<'a>

pub fn new_list(
arena: &'a Arena<Syntax<'a>>,
open_content: &str,
open_position: Vec<SingleLineSpan>,
children: Vec<&'a Syntax<'a>>,
close_content: &str,
close_position: Vec<SingleLineSpan>
) -> &'a Syntax<'a>

pub fn new_atom(
arena: &'a Arena<Syntax<'a>>,
position: Vec<SingleLineSpan>,
content: &str,
kind: AtomKind
) -> &'a Syntax<'a>

pub fn info(&self) -> &SyntaxInfo<'a>

pub fn next(&self) -> Option<&'a Syntax<'a>>

pub fn prev_is_contiguous(&self) -> bool

pub fn id(&self) -> u32

A unique ID of this syntax node. Every node is guaranteed to have a different value.

pub fn content_id(&self) -> u32

A content ID of this syntax node. Two nodes have the same content ID if they have the same content, regardless of position.

pub fn num_ancestors(&self) -> u32

pub fn first_line(&self) -> Option<LineNumber>

pub fn last_line(&self) -> Option<LineNumber>

pub fn change(&'a self) -> Option<ChangeKind<'a>>

pub fn set_change(&self, ck: ChangeKind<'a>)

pub fn set_change_deep(&self, ck: ChangeKind<'a>)

impl<'a> Syntax<'a>

fn first_line_span(&self) -> Option<SingleLineSpan>

fn last_line_span(&self) -> Option<SingleLineSpan>