Module difft::tree_sitter_parser

Load and configure parsers written with tree-sitter.

Structs

HighlightedNodeIds

TreeSitterConfig

Configuration for a tree-sitter parser.

Constants

OCAML_ATOM_NODES

Functions

all_syntaxes_from_cursor

Convert all the tree-sitter nodes at this level to difftastic syntax nodes.

atom_from_cursor

Convert the tree-sitter node at cursor to a difftastic atom.

child_tokens

find_delim_positions

Are any of the children of the node at cursor delimiters? Return their indexes if so.

from_language

list_from_cursor

Convert the tree-sitter node at cursor to a difftastic list node.

parse

Parse src with tree-sitter and convert to difftastic Syntax.

parse_to_tree

Parse src with tree-sitter.

print_cursor

print_tree

syntax_from_cursor

Convert the tree-sitter node at cursor to a difftastic syntax node.

tree_highlights

Calculate which tree-sitter node IDs should have which syntax highlighting.

tree_sitter_bash⚠

tree_sitter_c⚠

tree_sitter_c_sharp⚠

tree_sitter_clojure⚠

tree_sitter_commonlisp⚠

tree_sitter_cpp⚠

tree_sitter_css⚠

tree_sitter_dart⚠

tree_sitter_elisp⚠

tree_sitter_elixir⚠

tree_sitter_elm⚠

tree_sitter_elvish⚠

tree_sitter_gleam⚠

tree_sitter_go⚠

tree_sitter_haskell⚠

tree_sitter_hcl⚠

tree_sitter_janet_simple⚠

tree_sitter_java⚠

tree_sitter_javascript⚠

tree_sitter_json⚠

tree_sitter_kotlin⚠

tree_sitter_lua⚠

tree_sitter_nix⚠

tree_sitter_ocaml⚠

tree_sitter_ocaml_interface⚠

tree_sitter_perl⚠

tree_sitter_php⚠

tree_sitter_python⚠

tree_sitter_ruby⚠

tree_sitter_rust⚠

tree_sitter_scala⚠

tree_sitter_swift⚠

tree_sitter_toml⚠

tree_sitter_tsx⚠

tree_sitter_typescript⚠

tree_sitter_yaml⚠

tree_sitter_zig⚠