Crate difft

Difftastic is a syntactic diff tool.

For usage instructions and advice on contributing, see the manual.

Modules

constants

diff

display

exit_codes

files

File reading utilities.

line_parser

A fallback “parser” for plain text.

lines

Manipulate lines of text and groups of lines.

options

CLI option parsing.

parse

positions

Represents positions within a string.

summary

Data types summarising the result of diffing content.

Statics

GLOBAL

The global allocator used by difftastic.

Functions

check_only_text

diff_directories

Given two directories that contain the files, compare them pairwise. Returns an iterator, so we can print results incrementally.

diff_file

Print a diff between two files.

diff_file_content

format_num_bytes

main

The entrypoint.

print_diff_result

reset_sigpipe

Terminate the process if we get SIGPIPE.