Module difft::diff

Modules

changes 🔒

Data types that track the change state for syntax nodes.

dijkstra 🔒

Implements Dijkstra’s algorithm for shortest path, to find an optimal and readable diff between two ASTs.

graph 🔒

A graph representation for computing tree diffs.

myers_diff 🔒

A fast diff for linear content, using Myer’s diff algorithm.

sliders 🔒

Prefer contiguous novel nodes on the same line.

stack 🔒

unchanged 🔒

Find nodes that are obviously unchanged, so we can run the main diff on smaller inputs.