Module difft::diff::unchanged

Find nodes that are obviously unchanged, so we can run the main diff on smaller inputs.

Structs

EqOnFirstItem

Enums

ChangeState

Constants

MOSTLY_UNCHANGED_MIN_COMMON_CHILDREN

TINY_TREE_THRESHOLD

Functions

as_singleton_list_children

count_common_unique

Count the nodes in lhs that are unique to the LHS input, but are also present in rhs.

count_unique_subtrees

find_all_unique_content_ids

find_unique_content_ids

is_mostly_unchanged_list

Return true if both nodes are lists with same delimiters and have the same start and end children.

mark_unchanged

Set ChangeKind on nodes that are obviously unchanged, and return a vec of pairs that need proper diffing.

shrink_unchanged_at_ends

Skip syntax nodes at the beginning or end that are obviously unchanged.

shrink_unchanged_delimiters

If we’re comparing two lists that have the same delimiters, mark the delimiters as unchanged and return the children.

split_mostly_unchanged_toplevel

Split out top-level lists that are largely the same.

split_unchanged

split_unchanged_singleton_list

split_unchanged_toplevel

Mark top-level nodes as unchanged if they have exactly the same content on both sides.