Module difft::display::hunks

Calculating which modified lines should be displayed together.

Structs

Hunk

A hunk represents a series of modified lines that are displayed together.

Constants

MAX_DISTANCE

The maximum number of lines that may occur between changed lines in a hunk.

Functions

either_side_equal

enforce_increasing

Given a slice of line pairs that are ordered as far as possible, but may refer to earlier lines, return line pairs that are monotonically increasing. If an earlier line occurs after a later line, replace that line with None.

extract_lines

fill_between

find_novel_lines

lines_are_close

lines_to_hunks

Split lines into hunks.

matched_lines_indexes_for_hunk

matched_novel_lines

matched_pos_to_hunks

merge_adjacent

next_opposite

novel_section_in_order

Given a sequence of novel MatchedPos values in a section between two unchanged MatchedPos values, return them in an order suited for displaying.

sorted_novel_positions

Return a vec of novel MatchedPos values in an order suited for displaying.