Module difft::display::hunks

Calculating which modified lines should be displayed together.

Structs

Hunk ğŸ”’

A hunk represents a series of modified lines that are displayed together.

Constants

MAX_DISTANCE ğŸ”’

The maximum number of lines that may occur between changed lines in a hunk.

Functions

either_side_equal ğŸ”’

enforce_increasing ğŸ”’

Given a slice of line pairs that are ordered as far as possible, but may refer to earlier lines, return line pairs that are monotonically increasing. If an earlier line occurs after a later line, replace that line with None.

extract_lines ğŸ”’

fill_between ğŸ”’

find_novel_lines ğŸ”’

lines_are_close ğŸ”’

lines_to_hunks ğŸ”’

Split lines into hunks.

matched_lines_indexes_for_hunk ğŸ”’

Given a set of matched lines between the LHS and RHS, return the start and end indexes in matched_lines that should be displayed for hunk.

matched_novel_lines ğŸ”’

matched_pos_to_hunks ğŸ”’

merge_adjacent ğŸ”’

If any hunks are sufficiently close that they have overlapping lines, merge them.

next_opposite ğŸ”’

novel_section_in_order ğŸ”’

Given a sequence of novel MatchedPos values in a section between two unchanged MatchedPos values, return them in an order suited for displaying.

sorted_novel_positions ğŸ”’

Return a vec of novel MatchedPos values in an order suited for displaying.