Module difft::display::style

Apply colours and styling to strings.

Enums

BackgroundColor ğŸ”’

Functions

apply_colors ğŸ”’

apply_header_color ğŸ”’

apply_line ğŸ”’

Return a copy of line with styles applied to all the spans specified.

apply_line_number_color ğŸ”’

byte_offset_for_width ğŸ”’

Find the largest byte offset in s that gives the longest starting substring whose display width does not exceed width.

color_positions ğŸ”’

group_by_line ğŸ”’

header ğŸ”’

novel_style ğŸ”’

print_warning ğŸ”’

Style s as a warning and write to stderr.

replace_tabs ğŸ”’

Return a copy of src with all the tab characters replaced by tab_width strings.

split_and_apply ğŸ”’

Split line (from the source code) into multiple lines of max_len (i.e. word wrapping), and apply styles to each part according to its original position in line.

split_string_by_width ğŸ”’

Split a string into parts whose display length does not exceed max_width.

style_lines ğŸ”’

Apply the Styles to the spans specified. Return a vec of the styled strings, including trailing newlines.

substring_by_byte ğŸ”’

substring_by_byte_replace_tabs ğŸ”’

width_respecting_tabs ğŸ”’