Module difft::display::style

Apply colours and styling to strings.

Enums

BackgroundColor

Functions

apply_colors

apply_header_color

apply_line

Return a copy of line with styles applied to all the spans specified.

apply_line_number_color

byte_offset_for_width

Find the largest byte offset in s that gives the longest starting substring whose display width does not exceed width.

color_positions

group_by_line

header

novel_style

replace_tabs

Return a copy of src with all the tab characters replaced by tab_width strings.

split_and_apply

Split line (from the source code) into multiple lines of max_len (i.e. word wrapping), and apply styles to each part according to its original position in line.

split_string_by_width

Split a string into parts whose display length does not exceed max_width.

style_lines

Apply the Styles to the spans specified. Return a vec of the styled strings, including trailing newlines.

substring_by_byte

substring_by_byte_replace_tabs

width_respecting_tabs