Module difft::parse::guess_language

Guess which programming language a file is written in.

This is heavily based on GitHub’s linguist, particularly its languages.yml.

Difftastic does not reuse languages.yml directly. Linguist has a larger set of language detection strategies.

Structs

LanguageIter

An iterator over the variants of Self

Enums

Language

Languages supported by difftastic. Each language here has a corresponding tree-sitter parser.

LanguageOverride

Functions

from_emacs_mode_header

Try to guess the language based on an Emacs mode comment at the beginning of the file.

from_glob

from_shebang

Try to guess the language based on a shebang present in the source.

guess

language_globs

File globs that identify languages based on the file path.

language_name

The language name shown to the user.

language_override_from_name

If there is a language called name (comparing case insensitively), return it. Treat "text" as an additional option.

looks_like_hacklang