Module difft::parse::syntax

Syntax tree definitions with change metadata.

Structs

MatchedPos ğŸ”’

SyntaxInfo ğŸ”’

Fields that are common to both Syntax::List and Syntax::Atom.

Enums

AtomKind ğŸ”’

MatchKind ğŸ”’

A matched token (an atom, a delimiter, or a comment word).

StringKind ğŸ”’

Different types of strings. We want to diff these the same way, but highlight them differently.

Syntax ğŸ”’

TokenKind ğŸ”’

Unlike atoms, tokens can be delimiters like {.

Functions

change_positions ğŸ”’

Walk nodes and return a vec of all the changed positions.

change_positions_ ğŸ”’

comment_positions ğŸ”’

dbg_pos ğŸ”’

filter_empty_ends ğŸ”’

Skip line spans at the beginning or end that have zero width.

find_nodes_with_unique_content ğŸ”’

Assumes that set_content_id has already run.

has_common_words ğŸ”’

Are there sufficient common words that we should only highlight individual changed words?

init_all_info ğŸ”’

Initialise all the fields in SyntaxInfo.

init_info ğŸ”’

init_info_on_side ğŸ”’

Set all the SyntaxInfo values for all the roots on a single side (LHS or RHS).

init_next_prev ğŸ”’

set_content_id ğŸ”’

set_content_is_unique ğŸ”’

set_content_is_unique_from_counts ğŸ”’

set_next_sibling ğŸ”’

set_num_after ğŸ”’

set_num_ancestors ğŸ”’

set_parent ğŸ”’

set_prev ğŸ”’

For every syntax node in the tree, mark the previous node according to a preorder traversal.

set_prev_sibling ğŸ”’

set_unique_id ğŸ”’

split_atom_words ğŸ”’

Given the text content from a comment or strings, split it into MatchedPos values for the novel and unchanged words.

zip_pad_shorter ğŸ”’

zip_repeat_shorter ğŸ”’

Zip lhs with rhs, but repeat the last item from the shorter slice.

Type Definitions

ContentKey ğŸ”’

SyntaxId ğŸ”’