Module difft::parse::tree_sitter_parser

Load and configure parsers written with tree-sitter.

Structs

ExceededByteLimit ğŸ”’

ExceededParseErrorLimit ğŸ”’

HighlightedNodeIds ğŸ”’

TreeSitterConfig ğŸ”’

Configuration for a tree-sitter parser.

TreeSitterSubLanguage ğŸ”’

A language may contain certain nodes that are in other languages and should be parsed as such (e.g. HTML <script> nodes containing JavaScript). This struct describes how to identify such nodes, and what languages we should parse them as.

Constants

OCAML_ATOM_NODES ğŸ”’

Functions

all_syntaxes_from_cursor ğŸ”’

Convert all the tree-sitter nodes at this level to difftastic syntax nodes.

atom_from_cursor ğŸ”’

Convert the tree-sitter node at cursor to a difftastic atom.

child_tokens ğŸ”’

comment_positions ğŸ”’

find_delim_positions ğŸ”’

Are any of the children of the node at cursor delimiters? Return their indexes if so.

from_language ğŸ”’

list_from_cursor ğŸ”’

Convert the tree-sitter node at cursor to a difftastic list node.

parse ğŸ”’

Parse src with tree-sitter and convert to difftastic Syntax.

parse_subtrees ğŸ”’

Find any nodes that can be parsed as other languages (e.g. JavaScript embedded in HTML), and return a map of their node IDs mapped to parsed trees. Every time we see such a node, we will ignore it and recurse into the root node of the given tree instead.

print_cursor ğŸ”’

print_tree ğŸ”’

syntax_from_cursor ğŸ”’

Convert the tree-sitter node at cursor to a difftastic syntax node.

to_syntax ğŸ”’

to_syntax_with_limit ğŸ”’

to_tree ğŸ”’

Parse src with tree-sitter.

to_tree_with_limit ğŸ”’

tree_highlights ğŸ”’

Calculate which tree-sitter node IDs should have which syntax highlighting.

tree_sitter_commonlisp ğŸ”’ âš

tree_sitter_elvish ğŸ”’ âš

tree_sitter_hack ğŸ”’ âš

tree_sitter_hare ğŸ”’ âš

tree_sitter_janet_simple ğŸ”’ âš

tree_sitter_kotlin ğŸ”’ âš

tree_sitter_latex ğŸ”’ âš

tree_sitter_scss ğŸ”’ âš

tree_sitter_smali ğŸ”’ âš

tree_sitter_vhdl ğŸ”’ âš