Module difft::parse::tree_sitter_parser

Load and configure parsers written with tree-sitter.

Structs

ExceededByteLimit ğŸ”’

ExceededParseErrorLimit ğŸ”’

HighlightedNodeIds ğŸ”’

TreeSitterConfig ğŸ”’

Configuration for a tree-sitter parser.

TreeSitterSubLanguage ğŸ”’

A language may contain certain nodes that are in other languages and should be parsed as such (e.g. HTML <script> nodes containing JavaScript). This contains how to identify such nodes, and what languages we should parse them as.

Constants

OCAML_ATOM_NODES ğŸ”’

Functions

all_syntaxes_from_cursor ğŸ”’

Convert all the tree-sitter nodes at this level to difftastic syntax nodes.

atom_from_cursor ğŸ”’

Convert the tree-sitter node at cursor to a difftastic atom.

child_tokens ğŸ”’

comment_positions ğŸ”’

find_delim_positions ğŸ”’

Are any of the children of the node at cursor delimiters? Return their indexes if so.

from_language ğŸ”’

list_from_cursor ğŸ”’

Convert the tree-sitter node at cursor to a difftastic list node.

parse ğŸ”’

Parse src with tree-sitter and convert to difftastic Syntax.

parse_subtrees ğŸ”’

Find any nodes that can be parsed as other languages (e.g. JavaScript embedded in HTML), and return a map of their node IDs mapped to parsed trees. Every time we see such a node, we will ignore it and recurse into the root node of the given tree instead.

print_cursor ğŸ”’

print_tree ğŸ”’

syntax_from_cursor ğŸ”’

Convert the tree-sitter node at cursor to a difftastic syntax node.

to_syntax ğŸ”’

to_syntax_with_limit ğŸ”’

to_tree ğŸ”’

Parse src with tree-sitter.

to_tree_with_limit ğŸ”’

tree_highlights ğŸ”’

Calculate which tree-sitter node IDs should have which syntax highlighting.

tree_sitter_ada ğŸ”’ âš

tree_sitter_apex ğŸ”’ âš

tree_sitter_bash ğŸ”’ âš

tree_sitter_c ğŸ”’ âš

tree_sitter_c_sharp ğŸ”’ âš

tree_sitter_clojure ğŸ”’ âš

tree_sitter_cmake ğŸ”’ âš

tree_sitter_commonlisp ğŸ”’ âš

tree_sitter_cpp ğŸ”’ âš

tree_sitter_css ğŸ”’ âš

tree_sitter_dart ğŸ”’ âš

tree_sitter_elisp ğŸ”’ âš

tree_sitter_elixir ğŸ”’ âš

tree_sitter_elm ğŸ”’ âš

tree_sitter_elvish ğŸ”’ âš

tree_sitter_erlang ğŸ”’ âš

tree_sitter_gleam ğŸ”’ âš

tree_sitter_go ğŸ”’ âš

tree_sitter_hack ğŸ”’ âš

tree_sitter_hare ğŸ”’ âš

tree_sitter_haskell ğŸ”’ âš

tree_sitter_hcl ğŸ”’ âš

tree_sitter_html ğŸ”’ âš

tree_sitter_janet_simple ğŸ”’ âš

tree_sitter_java ğŸ”’ âš

tree_sitter_javascript ğŸ”’ âš

tree_sitter_json ğŸ”’ âš

tree_sitter_julia ğŸ”’ âš

tree_sitter_kotlin ğŸ”’ âš

tree_sitter_latex ğŸ”’ âš

tree_sitter_lua ğŸ”’ âš

tree_sitter_make ğŸ”’ âš

tree_sitter_newick ğŸ”’ âš

tree_sitter_nix ğŸ”’ âš

tree_sitter_objc ğŸ”’ âš

tree_sitter_ocaml ğŸ”’ âš

tree_sitter_ocaml_interface ğŸ”’ âš

tree_sitter_pascal ğŸ”’ âš

tree_sitter_perl ğŸ”’ âš

tree_sitter_php ğŸ”’ âš

tree_sitter_python ğŸ”’ âš

tree_sitter_qmljs ğŸ”’ âš

tree_sitter_r ğŸ”’ âš

tree_sitter_racket ğŸ”’ âš

tree_sitter_ruby ğŸ”’ âš

tree_sitter_rust ğŸ”’ âš

tree_sitter_scala ğŸ”’ âš

tree_sitter_scheme ğŸ”’ âš

tree_sitter_scss ğŸ”’ âš

tree_sitter_smali ğŸ”’ âš

tree_sitter_solidity ğŸ”’ âš

tree_sitter_sql ğŸ”’ âš

tree_sitter_swift ğŸ”’ âš

tree_sitter_toml ğŸ”’ âš

tree_sitter_tsx ğŸ”’ âš

tree_sitter_typescript ğŸ”’ âš

tree_sitter_vhdl ğŸ”’ âš

tree_sitter_xml ğŸ”’ âš

tree_sitter_yaml ğŸ”’ âš

tree_sitter_zig ğŸ”’ âš