Module difft::parse::tree_sitter_parser

Load and configure parsers written with tree-sitter.

Structs

ExceededByteLimit

ExceededParseErrorLimit

HighlightedNodeIds

TreeSitterConfig

Configuration for a tree-sitter parser.

TreeSitterSubLanguage

A language may contain certain nodes that are in other languages and should be parsed as such (e.g. HTML